pyproject.toml
[project]
name = "ticket-triage-openenv"
version = "0.1.0"
description = "Support ticket triage OpenEnv for Meta PyTorch hackathon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "requests>=2.25.0",
    "openai>=1.40.0",
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ticket_triage_openenv", "server"]
